Structural Rehabilitation
Structural Rehabilitation and Seismic Retrofitting Optimization with Composite Column Section in openseespy:
In this Analysis, by defining the Concrete Structure and Nonlinear Dynamic Analysis and specifying the Demand Displacement, we look for the optimal Plate Thickness and the Plate Length with Newton-raphson Method.
F = (Max. Abs. PGD) - Demand Displacement
This Method is good for Structural Rehabilitation and Seismic Retrofitting Optimization.
Newton-raphson Method is very Efficient and solves the optimization problem with good accuracy and speed for this kind complex Nonlinear Structural Analysis.

YouTube Video:
https://youtu.be/q_F5vDAAsSo