REBAR  DIAMETER AND CONCRETE SECTION
In this Analysis, by defining the Concrete Structure and the type of Nonlinear Dynamic Analysis and Nonlinear Pushover Analysis and specifying the Demand Displacement, we look for the optimal rebar diameter and the concrete section depth with Newton-raphson Method.
F = (Max. Abs. PGD) - Demand Displacement
Newton-raphson Method is very Efficient and solves the optimization problem with good accuracy and speed for this kind complex Nonlinear Structural Analysis

YouTube video:
https://youtu.be/K-9bMAPkoUo